Allow for custom data normalizers and similarity scorers

+++ src/main/java/com/gustavvy/ppefp/model/Candlestick.java
@@ -5,6 +5,6 @@ package com.gustavvy.ppefp.model;
  *
  * @author Gustav V. Y
  */
-public record Candlestick(long timestamp, float high, float low, float open, float close) {
+public record Candlestick(long timestamp, double high, double low, double open, double close) {
 
 }
